+++ packages/components/src/components/radio-button/radio-button.tsx
@@ -19,10 +19,10 @@ import {
   Prop,
 } from '@stencil/core';
 import classNames from 'classnames';
-import { emitEvent, generateUniqueId } from '../../utils/utils';
 import statusNote from '../../utils/status-note';
+import { emitEvent, generateUniqueId } from '../../utils/utils';
 
-interface InputChangeEventDetail {
+export interface InputChangeEventDetail {
   value: string | number | boolean | undefined | null;
 }
 
@@ -56,11 +56,11 @@ export class RadioButton {
   /** (optional) id or space separated list of ids of elements that provide or link to additional related information. */
   @Prop() ariaDetailsId?: string;
 
+  /** (optional) data-qa attribute for e2e testing */
+  @Prop() dataQa?: string;
+
   @Event({ eventName: 'scale-change' })
   scaleChange!: EventEmitter<InputChangeEventDetail>;
-  /** @deprecated in v3 in favor of kebab-case event names */
-  @Event({ eventName: 'scaleChange' })
-  scaleChangeLegacy!: EventEmitter<InputChangeEventDetail>;
 
   private readonly internalId = generateUniqueId();
 
@@ -89,12 +89,11 @@ export class RadioButton {
       if (this.checked) {
         this.uncheckSiblings();
       }
-      emitEvent(this, 'scaleChange', {
+      emitEvent(this, 'scale-change', {
         value: this.value == null ? this.value : this.value.toString(),
       });
     }
   };
-
   // We manually set `checked` to false on sibling <scale-radio-button> elements,
   // otherwise they stayed `checked` after being clicked once, forever.
   uncheckSiblings() {
@@ -145,6 +144,7 @@ export class RadioButton {
               ? ariaDescribedByAttr
               : {})}
             aria-details={this.ariaDetailsId}
+            {...(this.dataQa ? { 'data-qa': this.dataQa } : {})}
           />
           <label htmlFor={this.inputId}>{this.label}</label>
           {!!this.helperText && (